requirements.txt
Flask==3.1.3
python-dotenv==1.2.2
requests==2.33.1
urllib3==2.6.3
mysql-connector-python==8.3.0
bcrypt==4.1.2
cryptography==42.0.5
Werkzeug==3.1.3
